\name{MakeTable}
\alias{MakeTable}
\title{Make summary table}
\description{Summarized compound class information. Makes a table
  similar to Table 1 in the manuscript.
}
\usage{MakeTable(x = dolphin2011.meta)
}
\arguments{
  \item{x}{data frame of library metadata.}
}
\value{data frame
}
\details{\code{x} must have the format described in the help for
  \code{\link{LibraryReport}}.
}

