\name{PlotClasses}
\alias{PlotClasses}
\title{Plot 2D Chromatograms}
\description{Plot 1st and 2nd dimension retention times grouped by
  (1) halogenation, (2) three example compound classes, and (3) four
  unknown classes. Three panels. This is Figure 2 in the manuscript.
}
\usage{PlotClasses(lib = dolphin2011.meta)
}
\arguments{
  \item{lib}{data frame of library metadata.}
}
\value{Generates a PDF file.
}
\details{\code{lib} must have the format described in the help for \code{\link{LibraryReport}}.
}

