\name{SpecLibDolphin2011-package}
\alias{SpecLibDolphin2011-package}
\docType{package}
\title{
Mass Spectral Library Dolphin Blubber 2011
}
\description{
Contains the code and mass spectral library described in: Nontargeted Comprehensive Two-Dimensional Gas Chromatography/Time-of-Flight Mass Spectrometry Method and Software for Inventorying Persistent and Bioaccumulative Contaminants in Marine Environments, Environmental Science and Technology 2012, 46 (15), 8001-8008. DOI: 10.1021/es301139q. The sample is blubber from a common dolphin (Delphinus delphis) fatally stranded January 2006 in Orleans, Massachusetts. 

The article includes a description of
the study goals, sample, direct sample introduction method, GCxGC/TOF-MS
analytical method, the purpose and basic design of this software, and
results. The original PDF version of this library is supplementary
material to the article.

A PDF file containing user notes, SpecLibDolphin2011_instructions.pdf,
can be accessed by \code{help(package = "SpecLibDolphin2011")} then
clicking on browse directory, or by browsing the package source.
}
\keyword{ package }
