\name{SpecLibDolphin2011-package}
\alias{SpecLibDolphin2011-package}
\docType{package}
\title{
Mass Spectral Library Dolphin Blubber 2011
}
\description{
This package contains the code and data described in: A non-targeted
GCxGC/TOF-MS method and software for inventorying persistent and
bioaccumulative contaminants in marine environments, Environmental
Science and Technology, in press. This article includes a description of
the study goals, sample, direct sample introduction method, GCxGC/TOF-MS
analytical method, the purpose and basic design of this software, and
results. The original PDF version of this library is supplementary
material to the above manuscript.

A PDF file containing user notes, SpecLibDolphin2011_instructions.pdf,
can be accessed by \code{help(package = "SpecLibDolphin2011")} then
clicking on browse directory, or by browsing the package source. These user
notes are also included in the article supplementary material and the PDF version of the library on the OrgMassSpecR website.
}
\keyword{ package }